requirements.txt
torch
torchvision
numpy
matplotlib
dlib
opencv-python
tqdm
requests
pyyaml
pillow